changed coverageqcentry to have coverage as a number rather than string

+++ src/main/java/ca/corefacility/bioinformatics/irida/processing/impl/CoverageFileProcessor.java
@@ -1,7 +1,6 @@
 package ca.corefacility.bioinformatics.irida.processing.impl;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -11,7 +10,6 @@ import org.springframework.stereotype.Component;
 import ca.corefacility.bioinformatics.irida.model.joins.Join;
 import ca.corefacility.bioinformatics.irida.model.project.Project;
 import ca.corefacility.bioinformatics.irida.model.sample.CoverageQCEntry;
-import ca.corefacility.bioinformatics.irida.model.sample.QCEntry;
 import ca.corefacility.bioinformatics.irida.model.sample.Sample;
 import ca.corefacility.bioinformatics.irida.model.sample.SampleSequencingObjectJoin;
 import ca.corefacility.bioinformatics.irida.model.sequenceFile.SequencingObject;
@@ -58,9 +56,9 @@ public class CoverageFileProcessor implements FileProcessor {
 					return fastqc.getTotalBases();
 				}).sum();
 
-				Long coverage = totalBases / projectGenomeSize;
+				int coverage = (int) (totalBases / projectGenomeSize);
 				boolean positive = coverage >= requiredCoverage;
-				
+
 				// remove any existing coveratge entries
 				read.getQcEntries().stream().filter(q -> q instanceof CoverageQCEntry)
 						.forEach(q -> qcEntryRepository.delete(q));
